pyproject.toml
[tool.poetry]
name = "pokemon"
version = "0.1.0"
description = "Api rest to get pokemon information"
authors = ["kevinfcc26 <kevinfcc26@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
requests = "^2.32.3"
djangorestframework-simplejwt = "^5.3.1"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
flake8 = "^7.0.0"
black = "^24.4.2"
isort = "^5.13.2"
pylint = "^3.2.2"
pylint-django = "^2.5.5"
ipdb = "^0.13.13"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
addopts = "--cov --cov-report xml:coverage.xml -s -p no:warnings --durations=0 --durations-min=1"
python_files = "tests.py test_*.py *_tests.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
